test(c): split test_append.py integration -> integration/test_append.py (#40)
Moved TestPipelineAppendDispatch (3 tests) — pipeline + source-DB
mutation via INSERT INTO. TestLoadAppend (3 tests, destinations module
unit) stays for Wave D.
